requirements.txt
torch>=2.0.0
numpy>=1.23.0
pillow>=9.0.0
pytest>=7.4.0
